[enh] New getters on MEP model, integrating with TimePeriod manager
Until now, we had many getter property like:

    MEP.current_organizations
    MEP.old_organizations
    etc...

This lead to duplicate code. We fix this by using newly introduced
TimePeriod manager and queryset classes.
The old methods are now replaced with simple getters:

    MEP.get_organizations

(The get_ prefix is added to avoid overwriting the original field..)
We can use them with the manager helper methods like this:

    MEP.get_organizations.only_current()
    MEP.get_organizations.only_old()

This also work with the historical infos recently reintroduced in
Parltrack which sets a 9999-12-31 enddate on current mandates/groups...