[fix] On the "MEP that voted for" list, show the good group. fixes #274
A MEP who has been elected for more than one legislature can change
group, local party, or even country, but when details of a previous
legislature vote were displayed, the MEP data which was shown was the
current one.

This commit fixes this and prints the MEP's data at the date of the
vote.